requirements.txt
streamlit==1.31.0
pandas==2.1.1
plotly==5.18.0
numpy==1.26.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
supabase==1.2.0